Update 50-query-body-parameters.mdx
I have tried with single quotes and it didn't work. Should this be a valid JSON?
I guess not because for example here: https://scribe.knuckles.wtf/laravel/documenting/url-parameters example contains `Example: en` which is not wrapped in quotes.
Anyway, this would have helped me if I found it so it might help someone else.